Corections
Changed the filename from Securities to Titles and added the properties to add the text and images.
In Text, links were removed from the fonts of the letters.
In layouts, two Home and Desktop folders were created and we added the layouts file to each.
Edited the Index.astro to work with the home layouts.