Name change
Changed Text to Paragraph Body and all text names on the home page and component quantity and edited it to add any text.
Changed Titles to Header Body all tile names on the home page and component quantity.
Added some text to test it out, fix it up and put the correct path of components on the home page to import. Cleaned up the index and moved to the home page and specified props moved to component and started asking for props.